Backend/requirements.txt
# ========================================
# LEANMAKER BACKEND - DJANGO PURO + TYPESCRIPT
# ========================================
# Python 3.11+ | Django 4.2+ | Sin APIs REST
# ========================================

# Core Django
Django==4.2.7
django-cors-headers==4.3.1

# Database (SQL Server Azure)
django-mssql-backend==2.8.1
pyodbc==5.0.1

# Development & Testing
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
django-storages==1.14.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Handling
python-magic==0.4.27

# Date & JSON
python-dateutil==2.8.2
simplejson==3.18.4

# Email
django-anymail==10.2

# Caching
django-redis==5.3.0

# Type Checking (for development)
mypy==1.7.1
django-stubs==4.2.7

# Template Engine (for server-side rendering)
jinja2==3.1.2

# Static Files
whitenoise==6.6.0

# Security
django-ratelimit==4.1.0

# Logging
structlog==23.2.0 